Add files via upload
This program was just a fun-challenge to convert all chars in a string to be 2 alphabet letters ahead of what they currently are by modifying ASCII value. I.E. a->c, b->d, f->h. I did this in C++ as a beginning and wanted to reattempt in a new language.