Add files via upload
This is a program designed to take in an integer n, which will be the final number in our set. We can represent 2^n combinations of sets, and my program will output all the unique sets that can be made up to that number. E.g., 4  returns the individual last items in the list (4), then (5)
    #and after returns each possible pair for the last item, (4,5)
    #Sequence looks like this: 5,4, (4,5),3,(3,5),(3,4),(3,4,5), (2), (2,5)